Removed some old classes/views
