+++ tools/Titan.Tools.ManifestBuilder/Titan.Tools.ManifestBuilder.csproj
@@ -32,8 +32,8 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0-preview.7.22375.6" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Update="Views\ManifestView2.axaml.cs">
-      <DependentUpon>ManifestView2.axaml</DependentUpon>
+    <Compile Update="Views\Manifest\ManifestView.axaml.cs">
+      <DependentUpon>ManifestView.axaml</DependentUpon>
     </Compile>
     <Compile Update="Views\ProjectExplorerView.axaml.cs">
       <DependentUpon>ProjectExplorerView.axaml</DependentUpon>
