+++ tools/Titan.Tools.ManifestBuilder/ViewModels/ProjectExplorerViewModel.cs
@@ -7,6 +7,7 @@ using System.Windows.Input;
 using DynamicData.Binding;
 using ReactiveUI;
 using Titan.Tools.ManifestBuilder.Services;
+using Titan.Tools.ManifestBuilder.ViewModels.Manifest;
 using Titan.Tools.ManifestBuilder.Views.Dialogs;
 
 namespace Titan.Tools.ManifestBuilder.ViewModels;
@@ -21,14 +22,14 @@ public class ProjectExplorerViewModel : ViewModelBase
     }
     public bool HasManifests => Manifests.Count > 0;
     
-    public IObservableCollection<ManifestViewModel2> Manifests { get; } = new ObservableCollectionExtended<ManifestViewModel2>();
+    public IObservableCollection<ManifestViewModel> Manifests { get; } = new ObservableCollectionExtended<ManifestViewModel>();
 
-    private ManifestViewModel2? _selectedManifest;
+    private ManifestViewModel? _selectedManifest;
     private bool _projectLoaded;
     private readonly IManifestService _manifestService;
     private string? _projectPath;
 
-    public ManifestViewModel2? SelectedManifest
+    public ManifestViewModel? SelectedManifest
     {
         get => _selectedManifest;
         set => SetProperty(ref _selectedManifest, value);
@@ -48,7 +49,7 @@ public class ProjectExplorerViewModel : ViewModelBase
                 var createManifestResult = await _manifestService.CreateManifest(_projectPath!, name);
                 if (createManifestResult.Succeeded)
                 {
-                    var manifest = ManifestViewModel2.CreateFromManifest(createManifestResult.Data!);
+                    var manifest = ManifestViewModel.CreateFromManifest(createManifestResult.Data!);
                     Manifests.Add(manifest);
                     SelectedManifest = manifest;
                     this.RaisePropertyChanged(nameof(HasManifests));
@@ -67,7 +68,7 @@ public class ProjectExplorerViewModel : ViewModelBase
 
         if (manifests.Succeeded)
         {
-            var viewModels = manifests.Data!.Select(ManifestViewModel2.CreateFromManifest);
+            var viewModels = manifests.Data!.Select(ManifestViewModel.CreateFromManifest);
             Manifests.Load(viewModels);
         }
         SelectedManifest = Manifests.FirstOrDefault();
