+++ tools/Titan.Tools.ManifestBuilder/Views/Manifest/ManifestView.axaml
@@ -2,11 +2,11 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:viewModels="clr-namespace:Titan.Tools.ManifestBuilder.ViewModels"
+             xmlns:vm="clr-namespace:Titan.Tools.ManifestBuilder.ViewModels.Manifest"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="Titan.Tools.ManifestBuilder.Views.ManifestView2">
+             x:Class="Titan.Tools.ManifestBuilder.Views.Manifest.ManifestView">
   <Design.DataContext>
-    <viewModels:ManifestViewModel2/>
+    <vm:ManifestViewModel/>
   </Design.DataContext>
 
   <StackPanel>
