--- tools/Titan.Tools.ManifestBuilder/Views/ManifestView.axaml
@@ -1,33 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:viewModels="clr-namespace:Titan.Tools.ManifestBuilder.ViewModels"
-             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="Titan.Tools.ManifestBuilder.Views.ManifestView">
-
-  <Design.DataContext>
-    <viewModels:ManifestViewModel/>
-  </Design.DataContext>
-  <DockPanel LastChildFill="True">
-    <StackPanel DockPanel.Dock="Left">
-      <TextBlock Text="{Binding Title}"/>
-
-      <TextBlock IsVisible="{Binding Items.IsNotCompleted}">Loading....</TextBlock>
-      <TextBlock IsVisible="{Binding Items.IsFaulted}" Foreground="red">Error ocurred.</TextBlock>
-      <TextBlock IsVisible="{Binding Items.IsFaulted}" Foreground="red" Text="{Binding Items.ErrorMessage}" TextWrapping="Wrap"/>
-      <TextBlock IsVisible="{Binding HasError}" Foreground="purple" Text="{Binding Error}" TextWrapping="Wrap"/>
-      <TreeView IsVisible="{Binding Items.IsCompleted}"
-          Items="{Binding Items.Result}"
-                Width="400" Height="480"
-                HorizontalAlignment="Left"
-          SelectedItem="{Binding SelectedItem}">
-        <TreeView.ItemTemplate>
-          <TreeDataTemplate ItemsSource="{Binding Children}">
-            <TextBlock Text="{Binding Name}"/>
-          </TreeDataTemplate>
-        </TreeView.ItemTemplate>
-      </TreeView>
-    </StackPanel>
-  </DockPanel>
-</UserControl>
