--- tools/Titan.Tools.ManifestBuilder/Views/ManifestView2.axaml.cs
@@ -1,17 +1,0 @@
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace Titan.Tools.ManifestBuilder.Views;
-
-public partial class ManifestView2 : UserControl
-{
-    public ManifestView2()
-    {
-        InitializeComponent();
-    }
-
-    private void InitializeComponent()
-    {
-        AvaloniaXamlLoader.Load(this);
-    }
-}
